Add stress test for tenant-add
